feat(nav): implement main navigation components for user, workspaces, and main sections to enhance user experience and organization chore(sidebar): refactor sidebar to utilize new navigation components and remove deprecated hooks for cleaner codebase fix(projects): update project switcher to use new projects hook for dynamic project data retrieval
