+++ client/src/components/navs/hooks/useUser.tsx
@@ -1,0 +1,20 @@
+import {useMemo} from 'react';
+
+export interface UserDesc {
+    avatar: string;
+
+    email: string;
+
+    name: string;
+}
+
+export const useUser = (): UserDesc => {
+    return useMemo(
+        () => ({
+            name: 'shadcn',
+            email: 'm@example.com',
+            avatar: '/avatars/shadcn.jpg'
+        }),
+        []
+    );
+};
