+++ client/src/components/sidebar/AppSidebar.tsx
@@ -1,7 +1,4 @@
 import {ProjectSwitcher} from '@/components/projects/ProjectSwitcher';
-import {NavMain} from '@/components/shadcn/nav-main';
-import {NavProjects} from '@/components/shadcn/nav-projects';
-import {NavUser} from '@/components/shadcn/nav-user';
 import {
     Sidebar,
     SidebarContent,
@@ -9,162 +6,23 @@ import {
     SidebarHeader,
     SidebarRail
 } from '@/components/shadcn/ui/sidebar';
-import {
-    AudioWaveform,
-    BookOpen,
-    Bot,
-    Command,
-    Frame,
-    GalleryVerticalEnd,
-    Map,
-    PieChart,
-    Settings2,
-    SquareTerminal
-} from 'lucide-react';
+import {NavMain} from '@/components/navs/NavMain';
+import {NavUser} from '@/components/navs/NavUser';
+import {NavWorkspaces} from '@/components/navs/NavWorkspaces';
 import {ComponentProps, FC} from 'react';
-
-// This is sample data.
-const data = {
-    user: {
-        name: 'shadcn',
-        email: 'm@example.com',
-        avatar: '/avatars/shadcn.jpg'
-    },
-    teams: [
-        {
-            name: 'Acme Inc',
-            logo: GalleryVerticalEnd,
-            plan: 'Enterprise'
-        },
-        {
-            name: 'Acme Corp.',
-            logo: AudioWaveform,
-            plan: 'Startup'
-        },
-        {
-            name: 'Evil Corp.',
-            logo: Command,
-            plan: 'Free'
-        }
-    ],
-    navMain: [
-        {
-            title: 'Playground',
-            url: '#',
-            icon: SquareTerminal,
-            isActive: true,
-            items: [
-                {
-                    title: 'History',
-                    url: '#'
-                },
-                {
-                    title: 'Starred',
-                    url: '#'
-                },
-                {
-                    title: 'Settings',
-                    url: '#'
-                }
-            ]
-        },
-        {
-            title: 'Models',
-            url: '#',
-            icon: Bot,
-            items: [
-                {
-                    title: 'Genesis',
-                    url: '#'
-                },
-                {
-                    title: 'Explorer',
-                    url: '#'
-                },
-                {
-                    title: 'Quantum',
-                    url: '#'
-                }
-            ]
-        },
-        {
-            title: 'Documentation',
-            url: '#',
-            icon: BookOpen,
-            items: [
-                {
-                    title: 'Introduction',
-                    url: '#'
-                },
-                {
-                    title: 'Get Started',
-                    url: '#'
-                },
-                {
-                    title: 'Tutorials',
-                    url: '#'
-                },
-                {
-                    title: 'Changelog',
-                    url: '#'
-                }
-            ]
-        },
-        {
-            title: 'Settings',
-            url: '#',
-            icon: Settings2,
-            items: [
-                {
-                    title: 'General',
-                    url: '#'
-                },
-                {
-                    title: 'Team',
-                    url: '#'
-                },
-                {
-                    title: 'Billing',
-                    url: '#'
-                },
-                {
-                    title: 'Limits',
-                    url: '#'
-                }
-            ]
-        }
-    ],
-    projects: [
-        {
-            name: 'Design Engineering',
-            url: '#',
-            icon: Frame
-        },
-        {
-            name: 'Sales & Marketing',
-            url: '#',
-            icon: PieChart
-        },
-        {
-            name: 'Travel',
-            url: '#',
-            icon: Map
-        }
-    ]
-};
 
 export const AppSidebar: FC<ComponentProps<typeof Sidebar>> = (props) => {
     return (
         <Sidebar collapsible="icon" {...props}>
             <SidebarHeader>
-                <ProjectSwitcher projects={data.teams} />
+                <ProjectSwitcher />
             </SidebarHeader>
             <SidebarContent>
-                <NavMain items={data.navMain} />
-                <NavProjects projects={data.projects} />
+                <NavMain />
+                <NavWorkspaces />
             </SidebarContent>
             <SidebarFooter>
-                <NavUser user={data.user} />
+                <NavUser />
             </SidebarFooter>
             <SidebarRail />
         </Sidebar>
